Apply bug fixes and safety improvements (phase 6 batch 1)

+++ src/main/java/ch/technokrat/gecko/geckocircuits/circuit/CircuitSourceType.java
@@ -59,7 +59,7 @@ public enum CircuitSourceType {
         return _outputString;
     }
 
-    public double getNewID() {
+    public int getNewID() {
         return _newGeckoID;
     }
