+++ src/main/java/ch/technokrat/gecko/geckocircuits/circuit/ControlSourceType.java
@@ -54,7 +54,7 @@ public enum ControlSourceType {
         return _outputString;
     }
 
-    public double getNewID() {
+    public int getNewID() {
         return _newGeckoID;
     }
 }
